Extend routine `reset`, add state parsing, resetting of inner repositories, etc.

+++ proto/wtools/amid/l3/git/l1/Namespace.ss
@@ -4143,6 +4143,74 @@ configReset.defaults =
 
 */
 
+function _stateParse( state )
+{
+  let statesBegin = [ '#', '!' ];
+  let statesSpecial = [ 'working', 'staging', 'committed' ];
+  /*
+  https://neurathsboat.blog/post/git-intro/
+  https://git-scm.com/book/en/v2/Getting-Started-What-is-Git%3F
+  */
+
+  let result =
+  {
+    original : state,
+    value : state,
+    isSpecial : false
+  };
+
+  if( _.strBegins( state, 'HEAD' ) )
+  {
+    result.isSpecial = true;
+    return result;
+  }
+
+  if( _.strBegins( state, statesBegin ) )
+  {
+    result.isVersion = _.strBegins( state, statesBegin[ 0 ] );
+    result.isTag = _.strBegins( state, statesBegin[ 1 ] );
+    result.value = _.strRemoveBegin( state, statesBegin );
+
+    if( !result.isTag )
+    return result;
+
+    if( _.strEnds( result.value, '/' ) )
+    {
+      result.isTag = false;
+      result.value = _.strRemoveEnd( result.value, '/' );
+    }
+    else if( _.strHas( result.value, '/' ) )
+    {
+      result.isRemoteTag = true;
+      let r = _.strIsolateLeftOrNone( result.value, '/' );
+      _.sure
+      (
+        _.strDefined( r[ 0 ] ) && _.strDefined( r[ 2 ] ),
+        `Failed to parse state: ${result.original}, expects remote tag in format: "!remote/tag".`
+      );
+      result.remotePath = `:///${r[ 0 ]}`;
+      result.remote = `!${r[ 2 ]}`;
+    }
+
+    return result;
+  }
+
+  // if( !allowSpecial )
+  // throw _.err( `Expects state in one of formats:${statesBegin}, but got: ${state}` );
+
+  if( !_.longHas( statesSpecial, state ) )
+  {
+    debugger;
+    throw _.err( `Expects one of special states: ${statesSpecial}, but got: ${state}` );
+  }
+
+  result.isSpecial = true;
+
+  return result;
+}
+
+//
+
 function diff( o )
 {
   let self = this;
@@ -4160,8 +4228,8 @@ function diff( o )
 
   let ready = new _.Consequence().take( null );
   let result = Object.create( null );
-  let state1 = stateParse( o.state1 );
-  let state2 = stateParse( o.state2 ); /* qqq : ! aaa: special tags now work in both states */
+  let state1 = self._stateParse( o.state1 );
+  let state2 = self._stateParse( o.state2 ); /* qqq : ! aaa: special tags now work in both states */
 
   let start = _.process.starter
   ({
@@ -4190,74 +4258,6 @@ function diff( o )
   }
 
   return ready;
-
-  /* - */
-
-  function stateParse( state )
-  {
-    let statesBegin = [ '#', '!' ];
-    let statesSpecial = [ 'working', 'staging', 'committed' ];
-    /*
-    https://neurathsboat.blog/post/git-intro/
-    https://git-scm.com/book/en/v2/Getting-Started-What-is-Git%3F
-    */
-
-    let result =
-    {
-      original : state,
-      value : state,
-      isSpecial : false
-    };
-
-    if( _.strBegins( state, 'HEAD' ) )
-    {
-      result.isSpecial = true;
-      return result;
-    }
-
-    if( _.strBegins( state, statesBegin ) )
-    {
-      result.isVersion = _.strBegins( state, statesBegin[ 0 ] );
-      result.isTag = _.strBegins( state, statesBegin[ 1 ] );
-      result.value = _.strRemoveBegin( state, statesBegin );
-
-      if( !result.isTag )
-      return result;
-
-      if( _.strEnds( result.value, '/' ) )
-      {
-        result.isTag = false;
-        result.value = _.strRemoveEnd( result.value, '/' );
-      }
-      else if( _.strHas( result.value, '/' ) )
-      {
-        result.isRemoteTag = true;
-        let r = _.strIsolateLeftOrNone( result.value, '/' );
-        _.sure
-        (
-          _.strDefined( r[ 0 ] ) && _.strDefined( r[ 2 ] ),
-          `Failed to parse state: ${result.original}, expects remote tag in format: "!remote/tag".`
-        );
-        result.remotePath = `:///${r[ 0 ]}`;
-        result.remote = `!${r[ 2 ]}`;
-      }
-
-      return result;
-    }
-
-    // if( !allowSpecial )
-    // throw _.err( `Expects state in one of formats:${statesBegin}, but got: ${state}` );
-
-    if( !_.longHas( statesSpecial, state ) )
-    {
-      debugger;
-      throw _.err( `Expects one of special states: ${statesSpecial}, but got: ${state}` );
-    }
-
-    result.isSpecial = true;
-
-    return result;
-  }
 
   /* - */
 
@@ -4535,21 +4535,40 @@ diff.defaults =
 
 function reset( o )
 {
-  let ready = new _.Consequence().take( null );
+  let self = this;
 
-  _.routineOptions( reset, arguments );
   _.assert( arguments.length === 1 );
+  _.routineOptions( reset, arguments );
+  _.assert( _.strDefined( o.state ) );
   _.assert( _.strDefined( o.localPath ) );
+  _.assert( _.longHas( [ null, 'all' ], o.preset ) );
 
-  if( o.removingUntracked === null )
-  o.removingUntracked = 1;
+  if( o.dry )
+  return;
+
+  if( o.preset === 'all' )
+  {
+    _.assert( o.state === 'committed', 'Preset `all` resets all changes to latest commit' );
+
+    let o2 =
+    {
+      removingUntracked : 1,
+      removingSubrepositories : 1,
+      removingIgnored : 1,
+    };
+    _.mapExtend( o, o2 );
+  }
+
+  /* */
+
+  let ready = new _.Consequence().take( null );
 
   let start = _.process.starter
   ({
     sync : 0,
     deasync : 0,
     outputCollecting : 1,
-    mode : 'spawn',
+    mode : 'shell',
     currentPath : o.localPath,
     throwingExitCode : 0,
     inputMirroring : 0,
@@ -4557,9 +4576,31 @@ function reset( o )
     ready
   });
 
+  /* */
+
+  let state = self._stateParse( o.state );
+
+  if( state.value === 'working' )
+  return;
+
+  if( state.isTag )
+  debugger;
+
+  if( state.value === 'committed' )
   start( `git reset --hard` );
+  else if( state.isVersion || state.isTag )
+  start( `git reset --hard ${ state.value }` );
+
   if( o.removingUntracked )
-  start( `git clean -df` );
+  {
+    let command = `git clean -df`
+    if( o.removingSubrepositories )
+    command += 'f';
+    if( o.removingIgnored )
+    command += 'x';
+
+    start( command );
+  }
 
   /* qqq : should be "git clean -dffx", but not by default */
   /* qqq : cover each option */
@@ -4574,89 +4615,13 @@ function reset( o )
 
 reset.defaults =
 {
+  state : 'committed', /* 'working', 'staged', 'committed' some commit or tag */
   localPath : null,
   preset : null, /*[ null, 'all' ]*/ /* qqq : implement and cover option */
   removingUntracked : 1,
   removingIgnored : 0, /* qqq : implement and cover option */
   removingSubrepositories : 1, /* qqq : implement and cover option. option -ffx of git command clean */
   dry : 0, /* qqq : implement and cover option */
-  sync : 1,
-}
-
-//
-
-function restore( o )
-{
-  let ready = new _.Consequence().take( null );
-
-  _.assert( arguments.length === 1 );
-
-  _.routineOptions( restore, o );
-  _.assert( _.strDefined( o.localPath ) );
-  _.assert( o.source === null || _.strDefined( o.source ) );
-  _.assert( o.conflict === 'merge' || o.conflict === 'diff3' );
-
-  let start = _.process.starter
-  ({
-    sync : 0,
-    deasync : 0,
-    outputCollecting : 1,
-    mode : 'shell',
-    currentPath : o.localPath,
-    throwingExitCode : o.throwing,
-    inputMirroring : 0,
-    outputPiping : 0,
-    ready
-  });
-
-  o.source = o.source ? `--source ${ o.source }` : '--source HEAD';
-  o.workTree = o.workTree ? '--worktree' : '';
-  o.staged = o.staged ? '--staged' : '';
-  o.overlay = o.overlay ? '--overlay' : '--no-overlay';
-  o.quiet = o.quiet ? '--quiet' : '--no-quiet';
-  o.conflict = `--conflict=${ o.conflict }`;
-  if( o.fromFile )
-  {
-    _.assert( _.path.is( o.fromFile ), 'Expects path to file with pathspecs.' );
-    _.assert( !o.pathspec, 'Expects single type of pathspecs.' );
-    o.fromFile = `--pathspec-from-file=${ o.fromFile }`;
-  }
-  else
-  {
-    o.pathspec = o.pathspec ? o.pathspec : ':/';
-  }
-
-  let command = `git restore ${ o.source } ${ o.workTree } ${ o.staged } ${ o.overlay } ${ o.quiet } ${ o.conflict }`;
-  if( o.fromFile )
-  command += ` ${ o.fromFile }`;
-  else
-  command += ` -- ${ o.pathspec }`;
-
-  debugger;
-  start( command );
-
-  if( o.sync )
-  {
-    ready.deasync();
-    return ready.sync();
-  }
-  return ready;
-}
-
-restore.defaults =
-{
-  workTree : 1,
-  staged : 1,
-  source : null,
-  quiet : 0,
-  overlay : 0,
-  conflict : 'merge',
-  fromFile : null,
-  pathspec : null,
-
-  localPath : null,
-  dry : 0,
-  throwing : 1,
   sync : 1,
 }
 
@@ -4911,9 +4876,9 @@ let Extension =
   configSave,
   configReset, /* qqq : implement routine _.git.configReset() */
 
-  reset,
-  restore,
+  _stateParse,
   diff,
+  reset,
   renormalize,
 
 }
